requirements.txt
Flask
flask-cors
pymongo
google-cloud-aiplatform
vertexai
Werkzeug
gunicorn
requests
python-dotenv
pandas